tests: refactor execution spec tests by consolidating fixtures and removing redundant files
